+++ org.csstudio.display.builder.model/src/org/csstudio/display/builder/model/properties/CommonWidgetProperties.java
@@ -340,27 +340,4 @@ public class CommonWidgetProperties
     /** Runtime 'connected': Are all PVs of the widget connected? */
     public static final WidgetPropertyDescriptor<Boolean> runtimeConnected =
         newBooleanPropertyDescriptor(WidgetPropertyCategory.RUNTIME, "connected", Messages.WidgetProperties_Connected);
-
-    /** Runtime 'insets': Container widget representations may set these. */
-    public static final WidgetPropertyDescriptor<int[]> runtimeInsets =
-        new WidgetPropertyDescriptor<int[]>(
-            WidgetPropertyCategory.RUNTIME, "insets", Messages.WidgetProperties_Insets)
-    {
-        @Override
-        public WidgetProperty<int[]> createProperty(final Widget widget,
-                                                    final int[] value)
-        {
-            return new RuntimeWidgetProperty<int[]>(this, widget, value)
-            {
-                @Override
-                public void setValueFromObject(final Object value) throws Exception
-                {
-                    if (value instanceof int[]  &&  ((int[]) value).length == 2)
-                        setValue((int[]) value);
-                    else
-                        throw new Exception("Need int[2], got " + value);
-                }
-            };
-        }
-    };
 }
